+++ lib/fastly/models/batch.rb
@@ -12,14 +12,22 @@ require 'date'
 require 'time'
 
 module Fastly
-  # Results from VCL linting
-  class ValidatorResult
-    attr_accessor :data
+  class Batch
+    # A descriptor for the response of the entire batch
+    attr_accessor :title
+
+    # If an error is present in any of the requests, this field will describe that error
+    attr_accessor :type
+
+    # Per-key errors which failed to parse, validate, or otherwise transmit
+    attr_accessor :errors
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'data' => :'data'
+        :'title' => :'title',
+        :'type' => :'type',
+        :'errors' => :'errors'
       }
     end
 
@@ -31,7 +39,9 @@ module Fastly
     # Attribute type mapping.
     def self.fastly_types
       {
-        :'data' => :'ValidatorResultData'
+        :'title' => :'String',
+        :'type' => :'String',
+        :'errors' => :'Array<BatchErrors>'
       }
     end
 
@@ -45,19 +55,29 @@ module Fastly
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::ValidatorResult` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::Batch` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::ValidatorResult`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::Batch`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'data')
-        self.data = attributes[:'data']
+      if attributes.key?(:'title')
+        self.title = attributes[:'title']
+      end
+
+      if attributes.key?(:'type')
+        self.type = attributes[:'type']
+      end
+
+      if attributes.key?(:'errors')
+        if (value = attributes[:'errors']).is_a?(Array)
+          self.errors = value
+        end
       end
     end
 
@@ -79,7 +99,9 @@ module Fastly
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          data == o.data
+          title == o.title &&
+          type == o.type &&
+          errors == o.errors
     end
 
     # @see the `==` method
@@ -91,7 +113,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [data].hash
+      [title, type, errors].hash
     end
 
     # Builds the object from hash