+++ lib/fastly/models/validator_result_data_attributes_messages.rb
@@ -12,14 +12,22 @@ require 'date'
 require 'time'
 
 module Fastly
-  # Results from VCL linting
-  class ValidatorResult
-    attr_accessor :data
+  class ValidatorResultDataAttributesMessages
+    attr_accessor :type
+
+    attr_accessor :warning
+
+    attr_accessor :message
+
+    attr_accessor :tokens
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'data' => :'data'
+        :'type' => :'type',
+        :'warning' => :'warning',
+        :'message' => :'message',
+        :'tokens' => :'tokens'
       }
     end
 
@@ -31,7 +39,10 @@ module Fastly
     # Attribute type mapping.
     def self.fastly_types
       {
-        :'data' => :'ValidatorResultData'
+        :'type' => :'String',
+        :'warning' => :'Boolean',
+        :'message' => :'String',
+        :'tokens' => :'Array<Hash<String, TokensAdditionalProps>>'
       }
     end
 
@@ -45,19 +56,33 @@ module Fastly
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::ValidatorResult` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::ValidatorResultDataAttributesMessages` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::ValidatorResult`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::ValidatorResultDataAttributesMessages`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'data')
-        self.data = attributes[:'data']
+      if attributes.key?(:'type')
+        self.type = attributes[:'type']
+      end
+
+      if attributes.key?(:'warning')
+        self.warning = attributes[:'warning']
+      end
+
+      if attributes.key?(:'message')
+        self.message = attributes[:'message']
+      end
+
+      if attributes.key?(:'tokens')
+        if (value = attributes[:'tokens']).is_a?(Array)
+          self.tokens = value
+        end
       end
     end
 
@@ -65,12 +90,32 @@ module Fastly
     # @return Array for valid properties with the reasons
     def list_invalid_properties
       invalid_properties = Array.new
+      if @type.nil?
+        invalid_properties.push('invalid value for "type", type cannot be nil.')
+      end
+
+      if @warning.nil?
+        invalid_properties.push('invalid value for "warning", warning cannot be nil.')
+      end
+
+      if @message.nil?
+        invalid_properties.push('invalid value for "message", message cannot be nil.')
+      end
+
+      if @tokens.nil?
+        invalid_properties.push('invalid value for "tokens", tokens cannot be nil.')
+      end
+
       invalid_properties
     end
 
     # Check to see if the all the properties in the model are valid
     # @return true if the model is valid
     def valid?
+      return false if @type.nil?
+      return false if @warning.nil?
+      return false if @message.nil?
+      return false if @tokens.nil?
       true
     end
 
@@ -79,7 +124,10 @@ module Fastly
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          data == o.data
+          type == o.type &&
+          warning == o.warning &&
+          message == o.message &&
+          tokens == o.tokens
     end
 
     # @see the `==` method
@@ -91,7 +139,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [data].hash
+      [type, warning, message, tokens].hash
     end
 
     # Builds the object from hash