+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v5.3.1](https://github.com/fastly/fastly-ruby/releases/tag/release/v5.3.1) (2023-07-05)
+
+**Bug fixes:**
+
+- fix(snippet): dynamic field switched from int to string.
+- fix(vcl): implement correct response models.
+
 ## [v5.3.0](https://github.com/fastly/fastly-ruby/releases/tag/release/v5.3.0) (2023-06-27)
 
 **Enhancements:**